tools: Add 23.05 patch generator
rtw88 support various Linux versions using KERNEL_VERSION.
OpenWrt is a little bit weird: 23.05 use a heavily patched 5.15 kernel with mac80211 6.1 backports, and the default KERNEL_VERSION checks do not work correctly.

With the OPENWRT flag we can fix the checks, and this script add the flag to all relevant files on a rtw88 tree folder.

With this, we can generate patch files with 'git diff'.